fix: change block_id type from i32 to i64

+++ migrations/rollup_state/20210626034404_l2block.sql
@@ -1,5 +1,5 @@
 CREATE TABLE l2block (
-    block_id integer PRIMARY KEY,
+    block_id BIGINT PRIMARY KEY,
     new_root VARCHAR(256) NOT NULL,
     witness jsonb NOT NULL,
     created_time TIMESTAMP(0) NOT NULL DEFAULT CURRENT_TIMESTAMP
